@@ -6,14 +6,11 @@
   </PropertyGroup>-->
 
   <PropertyGroup>
-	  <TargetFramework>net7.0</TargetFramework>
+	  <TargetFramework>net8.0</TargetFramework>
 	  <ImplicitUsings>enable</ImplicitUsings>
 	  <Nullable>disable</Nullable>
   </PropertyGroup>
 
-  <ItemGroup>
-    <Folder Include="LanguageExtensions\" />
-  </ItemGroup>
 
 
 </Project>
