@@ -6,7 +6,7 @@
 	</PropertyGroup>-->
 
 	<PropertyGroup>
-		<TargetFramework>net7.0</TargetFramework>
+		<TargetFramework>net8.0</TargetFramework>
 		<ImplicitUsings>enable</ImplicitUsings>
 		<Nullable>disable</Nullable>
 	</PropertyGroup>
@@ -16,10 +16,10 @@
 	</ItemGroup>
 
 	<ItemGroup>
-	  <Content Include="Scripts.sql" />
+	  <Content Include="Scripts\Scripts.sql" />
 	</ItemGroup>
 	<ItemGroup>
-		<PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="5.0.9" />
+		<PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="8.0.1" />
 	</ItemGroup>
 	<ItemGroup>
 	  <ProjectReference Include="..\BaseDataValidatorLibrary\BaseDataValidatorLibrary.csproj" />
