@@ -1,30 +1,26 @@
-﻿
-using System;
-using System.Threading.Tasks;
-using EntityFrameworkCoreLibrary.Data;
+﻿using EntityFrameworkCoreLibrary.Data;
 
-namespace EntityFrameworkCoreLibrary.Classes
+namespace EntityFrameworkCoreLibrary.Classes;
+
+public class CreateOperations
 {
-    public class CreateOperations
+    public static async Task<(bool success, Exception exception)> NewExampleDatabase()
     {
-        public static async Task<(bool success, Exception exception)> NewExampleDatabase()
+
+        try
+        {
+            await using var context = new ExampleContext();
+            await context.Database.EnsureDeletedAsync();
+            await context.Database.EnsureCreatedAsync();
+
+            return (true, null);
+
+        }
+        catch (Exception e)
         {
 
-            try
-            {
-                await using var context = new ExampleContext();
-                await context.Database.EnsureDeletedAsync();
-                await context.Database.EnsureCreatedAsync();
+            return (false, e);
 
-                return (true, null);
-
-            }
-            catch (Exception e)
-            {
-
-                return (false, e);
-
-            }
         }
     }
 }
