@@ -1,20 +1,19 @@
 ﻿using EntityFrameworkCoreLibrary.Models;
 using Microsoft.EntityFrameworkCore;
 
-namespace EntityFrameworkCoreLibrary.Data
+namespace EntityFrameworkCoreLibrary.Data;
+
+public class ExampleContext : DbContext
 {
-    public class ExampleContext : DbContext
+    public DbSet<Customer> Customers { get; set; }
+
+    /// <summary>
+    /// For production place in appsettings.json
+    /// </summary>
+    private static readonly string ConnectionString = "Server=(localdb)\\mssqllocaldb;Database=EF.DataAnnotations;Trusted_Connection=True";
+
+    protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
     {
-        public DbSet<Customer> Customers { get; set; }
-
-        /// <summary>
-        /// For production place in appsettings.json
-        /// </summary>
-        private static readonly string ConnectionString = "Server=(localdb)\\mssqllocaldb;Database=EF.DataAnnotations;Trusted_Connection=True";
-
-        protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
-        {
-            optionsBuilder.UseSqlServer(ConnectionString);
-        }
+        optionsBuilder.UseSqlServer(ConnectionString);
     }
 }
