@@ -1,38 +1,35 @@
-﻿using System;
+﻿
 using System.ComponentModel.DataAnnotations;
 using RulesLibrary.LanguageExtensions;
 
-namespace RulesLibrary.Classes
+namespace RulesLibrary.Classes;
+
+[AttributeUsage(AttributeTargets.Field | AttributeTargets.Property)]
+public class YearRangeAttribute : ValidationAttribute
 {
-
-    [AttributeUsage(AttributeTargets.Field | AttributeTargets.Property)]
-    public class YearRangeAttribute : ValidationAttribute
+    public YearRangeAttribute(int maximumYear)
     {
-        public YearRangeAttribute(int maximumYear)
-        {
-            MaximumYear = maximumYear;
-        }
-        /// <summary>
-        /// Maximum permitted year
-        /// </summary>
-        public int MaximumYear { get; }
-        /// <summary>
-        /// Minimum permitted year
-        /// </summary>
-        public int MinimumYear { get; set; }
-        public override string FormatErrorMessage(string name)
-        {
-            if (ErrorMessage is null && ErrorMessageResourceName is null)
-            {
-                ErrorMessage = "'Year {0}' and/or year '{1}' are invalid";
-            }
-
-            return $"{name} year must be between {MinimumYear} and {MaximumYear}";
-
-        }
-
-        public override bool IsValid(object sender) 
-            => ((DateTime)sender).Year.Between(MinimumYear, MaximumYear);
+        MaximumYear = maximumYear;
     }
-}
+    /// <summary>
+    /// Maximum permitted year
+    /// </summary>
+    public int MaximumYear { get; }
+    /// <summary>
+    /// Minimum permitted year
+    /// </summary>
+    public int MinimumYear { get; set; }
+    public override string FormatErrorMessage(string name)
+    {
+        if (ErrorMessage is null && ErrorMessageResourceName is null)
+        {
+            ErrorMessage = "'Year {0}' and/or year '{1}' are invalid";
+        }
 
+        return $"{name} year must be between {MinimumYear} and {MaximumYear}";
+
+    }
+
+    public override bool IsValid(object sender) 
+        => ((DateTime)sender).Year.Between(MinimumYear, MaximumYear);
+}
