@@ -1,11 +1,10 @@
-﻿namespace WindowsFormsLibrary.Classes
+﻿namespace WindowsFormsLibrary.Classes;
+
+public class ApplicationSettings
 {
-	public class ApplicationSettings
-	{
-		public bool ShowAgain { get; set; }
-		public string Heading { get; set; }
-		public string Text { get; set; }
-		public string Caption { get; set; }
-		public string VerificationText { get; set; }
-	}
+    public bool ShowAgain { get; set; }
+    public string Heading { get; set; }
+    public string Text { get; set; }
+    public string Caption { get; set; }
+    public string VerificationText { get; set; }
 }
