@@ -1,70 +1,69 @@
 ﻿using System.Runtime.InteropServices;
 using System.Text;
 
-namespace WindowsFormsLibrary.Classes
+namespace WindowsFormsLibrary.Classes;
+
+/// <summary>
+/// Win forms-How can I make MessageBox appear centered on MainForm?
+/// https://stackoverflow.com/questions/2576156/winforms-how-can-i-make-messagebox-appear-centered-on-mainform
+/// </summary>
+public class CenterWinDialog : IDisposable
 {
-    /// <summary>
-    /// Win forms-How can I make MessageBox appear centered on MainForm?
-    /// https://stackoverflow.com/questions/2576156/winforms-how-can-i-make-messagebox-appear-centered-on-mainform
-    /// </summary>
-    public class CenterWinDialog : IDisposable
+    private int _tries = 0;
+    private readonly Form _formOwner;
+
+    public CenterWinDialog(Form owner)
     {
-        private int _tries = 0;
-        private readonly Form _formOwner;
-
-        public CenterWinDialog(Form owner)
-        {
-            _formOwner = owner;
-            owner.BeginInvoke(new MethodInvoker(findDialog));
-        }
-
-        private void findDialog()
-        {
-            // Enumerate windows to find the message box
-            if (_tries < 0) return;
-            EnumThreadWndProc callback = new(checkWindow);
-            if (EnumThreadWindows(GetCurrentThreadId(), callback, IntPtr.Zero))
-            {
-                if (++_tries < 10) _formOwner.BeginInvoke(new MethodInvoker(findDialog));
-            }
-        }
-        private bool checkWindow(IntPtr hWnd, IntPtr lp)
-        {
-            // Checks if <hWnd> is a dialog
-            StringBuilder sb = new (260);
-            GetClassName(hWnd, sb, sb.Capacity);
-            if (sb.ToString() != "#32770") return true;
-
-            // Got it
-            Rectangle frmRect = new Rectangle(_formOwner.Location, _formOwner.Size);
-            RECT dlgRect;
-            GetWindowRect(hWnd, out dlgRect);
-
-            MoveWindow(hWnd,
-                frmRect.Left + (frmRect.Width - dlgRect.Right + dlgRect.Left) / 2,
-                frmRect.Top + (frmRect.Height - dlgRect.Bottom + dlgRect.Top) / 2,
-                dlgRect.Right - dlgRect.Left,
-                dlgRect.Bottom - dlgRect.Top, true);
-            return false;
-        }
-        public void Dispose()
-        {
-            _tries = -1;
-        }
-
-        #region P/Invoke declarations
-        private delegate bool EnumThreadWndProc(IntPtr hWnd, IntPtr lp);
-        [DllImport("user32.dll")]
-        private static extern bool EnumThreadWindows(int tid, EnumThreadWndProc callback, IntPtr lp);
-        [DllImport("kernel32.dll")]
-        private static extern int GetCurrentThreadId();
-        [DllImport("user32.dll")]
-        private static extern int GetClassName(IntPtr hWnd, StringBuilder buffer, int buflen);
-        [DllImport("user32.dll")]
-        private static extern bool GetWindowRect(IntPtr hWnd, out RECT rc);
-        [DllImport("user32.dll")]
-        private static extern bool MoveWindow(IntPtr hWnd, int x, int y, int w, int h, bool repaint);
-        private struct RECT { public int Left; public int Top; public int Right; public int Bottom; }
-        #endregion
+        _formOwner = owner;
+        owner.BeginInvoke(new MethodInvoker(findDialog));
     }
+
+    private void findDialog()
+    {
+        // Enumerate windows to find the message box
+        if (_tries < 0) return;
+        EnumThreadWndProc callback = new(checkWindow);
+        if (EnumThreadWindows(GetCurrentThreadId(), callback, IntPtr.Zero))
+        {
+            if (++_tries < 10) _formOwner.BeginInvoke(new MethodInvoker(findDialog));
+        }
+    }
+    private bool checkWindow(IntPtr hWnd, IntPtr lp)
+    {
+        // Checks if <hWnd> is a dialog
+        StringBuilder sb = new (260);
+        GetClassName(hWnd, sb, sb.Capacity);
+        if (sb.ToString() != "#32770") return true;
+
+        // Got it
+        Rectangle frmRect = new Rectangle(_formOwner.Location, _formOwner.Size);
+        RECT dlgRect;
+        GetWindowRect(hWnd, out dlgRect);
+
+        MoveWindow(hWnd,
+            frmRect.Left + (frmRect.Width - dlgRect.Right + dlgRect.Left) / 2,
+            frmRect.Top + (frmRect.Height - dlgRect.Bottom + dlgRect.Top) / 2,
+            dlgRect.Right - dlgRect.Left,
+            dlgRect.Bottom - dlgRect.Top, true);
+        return false;
+    }
+    public void Dispose()
+    {
+        _tries = -1;
+    }
+
+    #region P/Invoke declarations
+    private delegate bool EnumThreadWndProc(IntPtr hWnd, IntPtr lp);
+    [DllImport("user32.dll")]
+    private static extern bool EnumThreadWindows(int tid, EnumThreadWndProc callback, IntPtr lp);
+    [DllImport("kernel32.dll")]
+    private static extern int GetCurrentThreadId();
+    [DllImport("user32.dll")]
+    private static extern int GetClassName(IntPtr hWnd, StringBuilder buffer, int buflen);
+    [DllImport("user32.dll")]
+    private static extern bool GetWindowRect(IntPtr hWnd, out RECT rc);
+    [DllImport("user32.dll")]
+    private static extern bool MoveWindow(IntPtr hWnd, int x, int y, int w, int h, bool repaint);
+    private struct RECT { public int Left; public int Top; public int Right; public int Bottom; }
+    #endregion
 }
