@@ -1,4 +1,5 @@
 ﻿using System.Text.RegularExpressions;
+#pragma warning disable SYSLIB1045
 
 namespace BaseDataValidatorLibrary.LanguageExtensions
 {
@@ -13,17 +14,18 @@ namespace BaseDataValidatorLibrary.LanguageExtensions
             string.Join(" ", Regex.Matches(sender, @"([A-Z][a-z]+)")
                 .Select(m => m.Value));
 
-        private static readonly Regex Reg = new("([a-z,0-9](?=[A-Z])|[A-Z](?=[A-Z][a-z]))", RegexOptions.Compiled);
+        private static readonly Regex Reg = new("([a-z,0-9](?=[A-Z])|[A-Z](?=[A-Z][a-z]))", 
+            RegexOptions.Compiled);
 
         /// <summary>
         /// This splits up a string based on capital letters
         /// e.g. "MyAction" would become "My Action" and "My10Action" would become "My10 Action"
         /// </summary>
-        /// <param name="str"></param>
+        /// <param name="sender"></param>
         /// <returns></returns>
-        public static string SplitPascalCase(this string str)
+        public static string SplitPascalCase(this string sender)
         {
-            return Reg.Replace(str, "$1 ");
+            return Reg.Replace(sender, "$1 ");
         }
     }
 }
