@@ -19,11 +19,8 @@ namespace RulesLibrary.LanguageExtensions
         /// This splits up a string based on capital letters
         /// e.g. "MyAction" would become "My Action" and "My10Action" would become "My10 Action"
         /// </summary>
-        /// <param name="str"></param>
+        /// <param name="sender"></param>
         /// <returns></returns>
-        public static string SplitPascalCase(this string str)
-        {
-            return Reg.Replace(str, "$1 ");
-        }
+        public static string SplitPascalCase(this string sender) => Reg.Replace(sender, "$1 ");
     }
 }
